wasm exeuctor models refactored
