+++ pkg/models/utils/job.go
@@ -1,7 +1,7 @@
 package utils
 
 import (
-	wasm "github.com/bacalhau-project/bacalhau/pkg/executor/wasm/models"
+	"github.com/bacalhau-project/bacalhau/pkg/executor/wasm"
 	"github.com/bacalhau-project/bacalhau/pkg/models"
 	"github.com/rs/zerolog/log"
 	"github.com/samber/lo"
